refactor: keep UtilButton as separate component
